comparison of Float with String failed
Somewhere in the stack trace after running `boxen`:

```sh
Error: comparison of Float with String failed
/opt/boxen/homebrew/Library/Taps/boxen-brews/dnsmasq.rb:28:in `<='
/opt/boxen/homebrew/Library/Taps/boxen-brews/dnsmasq.rb:28:in `install'
/opt/boxen/homebrew/Library/Taps/boxen-brews/dnsmasq.rb:15
```

Upon further investigation, it appears the `MACOS_VERSION` constant was recently
changed from a `Float` to a `String`.

To prevent further issues, instead of assuming the `MACOS_VERSION` constant is
always going to be a `Float`, I suggest we actually coerce it into a `Float`
before comparison.