Fix inconsistent service name
When the new tld option was introduced, the name of the service became
tied to the TLD used. This is screwy, because "dev" is hard-coded as a
service name prefix in the boxen ruby gem.  We could conceivably add
more TLDs to our Dnsmasq setup, but we won't run a separate service for
each one. Also, we should stick with the convention observed in other
boxen-managed services where the service name is not tied to the
configuration of the service.